TOUT FONCTIONNE

+++ delconfigserv.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
 
-#echo "Attention, cette requête va complètement nettoyer votre machine. Voulez-vous continuer? (o/n)"
+echo "Attention, cette requête va complètement nettoyer votre machine. Voulez-vous continuer? (o/n)"
 
-#read answer
+read answer
 
-#if $answer = "o"; then
-#echo "WP"
+if [ $answer = "o" ]; then
 apt --purge autoremove proftpd-*
-#else
-#exit 1
-#fi
+
+else
+exit 1
+
+fi
